starter 去除

+++ flow-core/src/main/java/com/flowable/oa/core/entity/BizFile.java
@@ -28,10 +28,10 @@ public class BizFile implements Serializable, Cloneable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "biz_id")
-    private Integer bizId;
+    private Long bizId;
 
     @Column(length = 512, name = "task_name")
     private String taskName;
