+++ flow-core/src/main/java/com/flowable/oa/core/service/IProcessVariableService.java
@@ -1,32 +1,34 @@
 package com.flowable.oa.core.service;
 
-import java.util.List;
-
 import com.flowable.oa.core.entity.ProcessVariable;
 import com.flowable.oa.core.util.mybatis.IBaseService;
 import com.github.pagehelper.PageInfo;
+import java.util.List;
 import org.flowable.engine.repository.ProcessDefinition;
 
 /**
- * 流程属性模板业务，包括模板属性以及任务属性处理
+ * @author : yangqi
+ * @email : lukewei@mockuai.com
+ * @description : 流程属性模板业务，包括模板属性以及任务属性处理
+ * @since : 2020/2/26 9:25 下午
  */
 public interface IProcessVariableService extends IBaseService<ProcessVariable> {
 
-	/**
-	 * 删除
-	 * 
-	 * @param list
-	 * @
-	 */
-	void deleteVariable(List<Integer> list);
+    /**
+     * 删除
+     *
+     * @param list
+     * @
+     */
+    void deleteVariable(List<Long> list);
 
-	PageInfo<ProcessVariable> findProcessVariables(ProcessVariable variable, PageInfo<ProcessVariable> page);
+    PageInfo<ProcessVariable> findProcessVariables(ProcessVariable variable, PageInfo<ProcessVariable> page);
 
-	/**
-	 * 参数复制
-	 * 
-	 * @param oldPdf
-	 * @param newPdf
-	 */
-	void copyVariables(ProcessDefinition oldPdf, ProcessDefinition newPdf);
+    /**
+     * 参数复制
+     *
+     * @param oldPdf
+     * @param newPdf
+     */
+    void copyVariables(ProcessDefinition oldPdf, ProcessDefinition newPdf);
 }
